Setup and initialization notes:
This is a "learning" project.  The code itself will be open source however I CAN NOT release the textures files, as they exist at initialization, because they are property of https://gamedev.tv.

I may replace these files with my own, or with open source alternatives, as this project proceeds. <3